wip ops for custom bert
